Problem: stream_engine assumes options_t.type is a single byte, but its type is int
Solution: change options_t.type to int8_t

+++ src/options.hpp
@@ -103,7 +103,7 @@ struct options_t
     int tos;
 
     //  Socket type.
-    int type;
+    int8_t type;
 
     //  Linger time, in milliseconds.
     atomic_value_t linger;
